refactor: abstract out stats update logic to `StatsAggregator`
