+++ packages/db/index.ts
@@ -1,6 +1,10 @@
 import { PrismaClient } from "@prisma/client";
 
+import { StatsAggregator } from "./StatsAggregator";
+
 export * from "@prisma/client";
+
+export * from "./utils/dates";
 
 const globalForPrisma = globalThis as unknown as { prisma: PrismaClient };
 
@@ -12,5 +16,6 @@ export const prisma =
         ? ["query", "error", "warn"]
         : ["error"],
   });
+export const statsAggregator = new StatsAggregator(prisma);
 
 if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
