+++ packages/db/tsconfig.json
@@ -2,15 +2,8 @@
   "extends": "../../tsconfig.json",
   "include": [
     "index.ts",
-    "prisma/seed.ts",
-    "prisma/stats.ts",
-    "prisma/utils/dates.ts"
-  ],
-  "ts-node": {
-    // these options are overrides used only by ts-node
-    // same as the --compilerOptions flag and the TS_NODE_COMPILER_OPTIONS environment variable
-    "compilerOptions": {
-      "module": "commonjs"
-    }
-  }
+    "prisma/**/*.ts",
+    "utils/**/*.ts",
+    "StatsAggregator/**/*.ts"
+  ]
 }
